feat(client): use mini-client
Replace built-in client with standalone mini-client and mini-service-utils libraries.
