--- index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  startServer: require('./lib/server'),
-  getClient: require('./lib/client')
-}
