+++ lib/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  startServer: require('./server'),
+  getClient: require('mini-client')
+}
