+++ lib/server.js
@@ -1,6 +1,7 @@
 const {Server} = require('hapi')
-const {getLogger} = require('./utils')
-const {registerAsPlugins} = require('./utils/register')
+const {getLogger} = require('mini-service-utils')
+const {registerAsPlugins} = require('./register')
+const package = require('../package.json')
 
 /**
  * Configure and start service in a standalone Hapi Http server
@@ -25,24 +26,30 @@ module.exports = (opts = {}) => {
 
   const options = Object.assign({
     port: 3000,
+    name: package.name,
+    version: package.version,
     logger: getLogger(),
     serviceOpts: {},
-    services: require('./services')
+    services: []
   }, opts)
 
-  const {port, logger, serviceOpts, services} = options
+  const {port, name, version, logger, serviceOpts, services} = options
   logger.debug({port}, 'Configure server')
 
   const server = new Server()
   server.connection({port})
 
   return registerAsPlugins(server, services, serviceOpts, logger)
-    .then(exposed => {
+    .then(apis => {
       // list of exposed APIs, for clients
       server.route({
         method: 'GET',
         path: '/api/exposed',
-        handler: (req, reply) => reply(exposed)
+        handler: (req, reply) => reply({
+          name,
+          version,
+          apis
+        })
       })
     })
     .then(() => server.start())
